Split expr tree string helpers
